Add basic symbol table
The symbol table is hierarchical with nested symbol tables. Nested
symbol tables can either be procedures, functions (procedures that
have a type and return values), or other modules. This symbol
table is generated before the IR AST due to the way IPL code is
structured.

Currently it handles the basic cases of global scope; there is more
to be done.