Created nested symtables for nested scopes
The root symtable now contains children for each function/procedure
which have their own lexical scoping. This is accomplished by giving
the visitor an overridden function to also call an exit function so
that lexical scoping can be tracked in a stack. Because IPL cannot
have lexical scopes within lexical scopes the symbol tree should, for
the moment, always be a root node with child who do not have children.