Use customized transformer to create AST
Lark's transformer does not allow keeping track of state or exit
functions by default. The TreeToAstTransformer (a mild tautology)
accomplishes this by duplicating most of Lark's transformer with
important changes that allow for this.

A design decision made with these changes is to give declared
variables a default value which may be None if its not a Python
primitive type. From the viewpoint of a compiler pipeline it makes
more sense to try and resolve these later during an optimization
that resolves that case where some initialization is overwritten
without first being referenced somewhere (i.e., `Int a = 1; a = 2`).