Add more AST transformations
Added transformer callbacks for subscripts and attributes, as well as
functionality to keep an internal context stack. This is useful for
when we may have nested subscripts or attributes that reference
attribute names which might otherwise be raised as an undeclared
identifier.

Also added functionality to distinguish between the IPL constants
and undeclared names. IPL allows its constants to be redefined so
a check is done for this as well, giving precedence to the redfined
identifier if it exists.