Simplify dependency graph
Single-use visitors and transformers did not need to be abstracted
out into a separate namespaces. This makes the dependency graph much
cleaner and localizes code to the only place that it is and will be
used. This also sneaks in functionality to be able to dump the AST
as yaml.