Add For and While loop code generation
This required breaking the for_stmt into two separate aliased rules
for increasing and decreasing loops. It is also a prime target for
some semantic interpretation because it translates such loops directly
into Python `range` loops, which is not very Pythonic. It also
requires a bit of a hack to append a single increment or decrement
so that the loop bound is inclusive, as this is not the case in
Python with range loops.