Add Point codegen and defaults and start of arrays
IPL always give a point in the x, y, z planes, even if only two
coordinates are provided, and it defaults all values to be floats.
Therefore ensure this remains the case when converted to a Python
tuple.

This commit also sketches some ideas for arrays but it's clear
that a transformation/optimization step is needed before it reaches
the code generation stage. This is because arrays in IPL are
initialized with a bare declaration, no size given, and then arbitrary
elements are added directly to an index one by one. The arrays are
resized automatically as this is done. While we could just change
these subscripts into array extensions, the semantic analysis needed
to do this already puts us close enough to being able to do a more
proper Python list/numpy array initialization.